refactor: card creator & related logic
