+++ src/components/card-creator/card-validator.ts
